+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.4.5]
+### Changes
+- Add support for auto initiating accounts
+- Add support for limit the number of created account for a plugin type
+### Fixes
+
 ## [0.4.4]
 ### Changes
 - Fix issue with loading optional protected parameters